Added refactoring explanation in comments and readme, cleaned code

+++ src/main/java/com/example/payment/PaymentApi.java
@@ -1,5 +1,7 @@
 package com.example.payment;
 
+import java.math.BigDecimal;
+
 public interface PaymentApi {
-    PaymentApiResponse charge(String API_KEY, double amount) throws PaymentException;
+    PaymentApiResponse charge(String API_KEY, BigDecimal amount) throws PaymentException;
 }
